minigames/DuckAttack/requirements.txt
pygame==2.5.2
qrcode==7.4.2
Pillow==10.0.1